Cliente: Controller + JSPs

+++ src/java/br/ufpr/tads/facade/AtendimentoFacade.java
@@ -5,8 +5,10 @@
 package br.ufpr.tads.facade;
 
 import br.ufpr.tads.beans.Atendimento;
+import br.ufpr.tads.beans.TipoAtendimento;
 import br.ufpr.tads.dao.AtendimentoDAO;
 import br.ufpr.tads.dao.ConnectionFactory;
+import br.ufpr.tads.dao.TipoAtendimentoDAO;
 import br.ufpr.tads.exception.AtendimentoException;
 import br.ufpr.tads.exception.DAOException;
 import java.util.List;
@@ -18,8 +20,7 @@ import java.util.List;
 public class AtendimentoFacade {
 
     public static List<Atendimento> buscarTodosAtendimentosAbertos() throws AtendimentoException {
-        try {
-            ConnectionFactory connF = new ConnectionFactory();
+        try(ConnectionFactory connF = new ConnectionFactory()){
             AtendimentoDAO atendimentoDao = new AtendimentoDAO(connF.getConnection());
             
             List<Atendimento> atendimentos = atendimentoDao.buscarTodosAtendimentosAbertos();
@@ -30,8 +31,7 @@ public class AtendimentoFacade {
     }
 
     public static List<Atendimento> buscarTodosAtendimentos() throws AtendimentoException {
-        try {
-            ConnectionFactory connF = new ConnectionFactory();
+        try(ConnectionFactory connF = new ConnectionFactory()){
             AtendimentoDAO atendimentoDao = new AtendimentoDAO(connF.getConnection());
             
             List<Atendimento> atendimentos = atendimentoDao.buscarTodos();
@@ -42,8 +42,7 @@ public class AtendimentoFacade {
     }
 
     public static List<Atendimento> buscarAtendimentosPorCliente(int clienteId) throws AtendimentoException {
-        try{
-            ConnectionFactory connF = new ConnectionFactory();
+        try(ConnectionFactory connF = new ConnectionFactory()){
             AtendimentoDAO atendimentoDao = new AtendimentoDAO(connF.getConnection());
             List<Atendimento> atendimentos = atendimentoDao.buscarAtendimentosCliente(clienteId);
             return atendimentos;
@@ -53,8 +52,7 @@ public class AtendimentoFacade {
     }
 
     public static Atendimento buscarAtendimentoPorId(int atendimentoId) throws AtendimentoException {
-        try{
-            ConnectionFactory connF = new ConnectionFactory();
+        try(ConnectionFactory connF = new ConnectionFactory()){
             AtendimentoDAO atendimentoDao = new AtendimentoDAO(connF.getConnection());
             
             Atendimento atendimento = atendimentoDao.buscar(atendimentoId);
@@ -65,8 +63,7 @@ public class AtendimentoFacade {
     }
 
     public static void criarAtendimento(Atendimento atendimento) throws AtendimentoException {
-        try{
-            ConnectionFactory connF = new ConnectionFactory();
+        try(ConnectionFactory connF = new ConnectionFactory()){
             AtendimentoDAO atendimentoDao = new AtendimentoDAO(connF.getConnection());
             
             atendimentoDao.inserir(atendimento);
@@ -76,13 +73,36 @@ public class AtendimentoFacade {
     }
 
     public static void deletarAtendimento(int atendimentoId) throws AtendimentoException {
-        try{
-            ConnectionFactory connF = new ConnectionFactory();
+        try(ConnectionFactory connF = new ConnectionFactory()){
             AtendimentoDAO atendimentoDao = new AtendimentoDAO(connF.getConnection());
             
-            atendimentoDao.remover(atendimentoId);
+            Atendimento atendimento = AtendimentoFacade.buscarAtendimentoPorId(atendimentoId);
+            
+            // Verifica se o atendimento está em aberto, se estiver, remove
+            // do contrário, não faz nenhuma ação
+            if (atendimento.getSituacao().getSituacaoId() == 1) {
+                atendimentoDao.remover(atendimentoId);
+            }
         }catch(DAOException ex){
             throw new AtendimentoException("Erro deletando atendimento", ex);
+        }
+    }
+    
+    public static List<TipoAtendimento> buscarTodosTiposAtendimento() throws AtendimentoException {
+        try(ConnectionFactory connF = new ConnectionFactory()){
+            TipoAtendimentoDAO tipoAtendimentoDAO = new TipoAtendimentoDAO(connF.getConnection());
+            return tipoAtendimentoDAO.buscarTodos();
+        }catch(DAOException ex){
+            throw new AtendimentoException("Erro buscando todos os tipos de atendimento", ex);
+        }
+    }
+
+    public static TipoAtendimento buscarTipoAtendimento(int idTipoAtendimento) throws AtendimentoException {
+        try(ConnectionFactory connF = new ConnectionFactory()){
+            TipoAtendimentoDAO tipoAtendimentoDAO = new TipoAtendimentoDAO(connF.getConnection());
+            return tipoAtendimentoDAO.buscar(idTipoAtendimento);
+        }catch(DAOException ex){
+            throw new AtendimentoException("Erro buscando o tipo de atendimento id: " + idTipoAtendimento, ex);
         }
     }
 }
